+++ src/config.php
@@ -17,4 +17,9 @@ return [
 	// Additional field types that can be used as filter criteria.
 	'additionalFieldTypes' => [],
 
+	// Should CP Filters include filtering for Craft Commerce?
+	'includeCommerce' => false,
+
+	// Array of IDS for product types that should be filterable by this plugin.
+	'filterableProductTypeIds' => null
 ];