+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.2.0 - 2021-11-15
+### Added
+- Added support for Craft Commerce Products and Orders.
+
+### Fixed
+- Fixed lightswitch "onLabel" and "offLabel" support for pre-3.5.4 Craft versions.
+
+
 ## 1.1.3 - 2021-08-25
 
 ### Fixed